Rename timelineCollection to timeline-collection
